backend/requirements.txt
aenum==3.1.16
agno==2.0.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
apscheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
azure-ai-documentintelligence==1.0.2
azure-core==1.35.1
azure-identity==1.25.0
backoff==2.2.1
beautifulsoup4==4.13.5
bs4==0.0.2
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chonkie==1.3.0
click==8.3.0
cloudscraper==1.2.71
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
composio==0.8.16
composio-client==1.9.1
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
deepgram-sdk==4.8.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
edge-tts==7.2.3
email-validator==2.3.0
et-xmlfile==2.0.0
executing==2.2.1
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
feedparser==6.0.12
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.7.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.40.3
google-genai==1.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.15
lancedb==0.25.1
langfuse==3.5.1
lxml==6.0.2
lxml-html-clean==0.4.2
magika==0.6.2
mammoth==1.10.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mouseinfo==0.1.3
moviepy==2.2.1
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
mss==10.1.0
multidict==6.6.4
mypy-extensions==1.1.0
nest-asyncio==1.6.0
newspaper4k==0.9.3.1
nltk==3.9.1
nuitka==2.7.16
numpy==2.2.0
olefile==0.47
onnxruntime==1.20.1
openai==1.109.1
openinference-instrumentation==0.1.38
openinference-instrumentation-agno==0.1.15
openinference-semantic-conventions==0.1.21
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
ordered-set==4.1.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdfminer-six==20250506
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
proglog==0.1.12
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyaudio==0.2.14
pyautogui==0.9.54
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pydub==0.25.1
pygetwindow==0.0.9
pygments==2.19.2
pyjwt==2.10.1
pylance==0.37.0
pymsgbox==2.0.1
pyparsing==3.2.5
pypdf==6.1.0
pypdf2==3.0.1
pyperclip==1.11.0
pyreadline3==3.5.4
pyrect==0.2.0
pyscreeze==1.0.1
pysher==1.0.8
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.20
python-pptx==1.0.2
python-vlc==3.0.21203
pytimeparse==1.1.8
pytweening==1.2.0
pytz==2025.2
pywin32==311
pyyaml==6.0.3
pyzmq==27.1.0
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
screeninfo==0.8.1
sentry-sdk==2.39.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
speechrecognition==3.14.3
sqlalchemy==2.0.43
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tantivy==0.25.0
tenacity==9.1.2
tiktoken==0.11.0
tldextract==5.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wmi==1.5.1
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.20.1
youtube-transcript-api==1.0.3
yt-dlp==2025.9.23
zipp==3.23.0
zstandard==0.25.0
